Timetable v2: Extract filtering into a separate hook
